requirements.txt
Django==2.2.5
django-autoslug==1.9.6
django-cleanup==4.0.0
django-crispy-forms==1.7.2
Pillow==6.1.0
pytz==2019.2
sqlparse==0.3.0
